wip: fixing cypress type errors
