+++ package.json
@@ -11,7 +11,8 @@
   "sideEffects": false,
   "scripts": {
     "postinstall": "nx run-many build --target=tools-workspace,eslint-plugin",
-    "e2e": "yarn nx e2e platform-e2e",
+    "e2e": "yarn nx e2e platform-e2e -c dev",
+    "e2e:api": "nx serve platform-api -c dev",
     "cli": "scripts/nx/cli.sh",
     "docker:up": "scripts/docker/up.sh",
     "tunnel": "ngrok http --region=us --hostname=api.codelab-dev.ngrok.io 127.0.0.1:3000",
